Remove shared_context resource_behavior
I was in the process of updating this to use the webmock syntax for RSpec and
I realized the tests are not doing all that much.  It would be more work to dynamically
figure out how to build the URLs that we are expecting than it would take to just
test these individually in their classes.  Right now it doesn't build URLs and just asserting
that a request happened is weak in my opinion.