Adds :update to BlockScore::Base
Adds an :update method to BlockScore::Base to add more of an
ActiveRecord feel to the library.

In addition to the :update method, a protected :initialize_from
method was added. This method will reinitialize the attributes of a
given object from a provided values hash. The Stripe version of this
method updates the dynamically defined accessors and getters based on
the values have changed. I have not done that because I do not think the
value of an attribute changing is a positive signal as to whether or not
a getter/setter will be called for that attribute in the future.

Also adds VCR cassettes for the update action, and changes the relevant
spec from pending to active.

Targets #40.