Use attr_writer in lieu of :initialize_from
Swap `attr_writer :attributes` in lieu of `:initialize_from` at the
suggestion of Hound and CodeClimate. This is fine for now since the only
reinitialization involved in rewriting the hash values. However, in the
future, it may make more sense to bring back the `:initialize_from`
method or an equivalent if more action needs to be taken upon
reinitialization.

Add a quick example to comments at the top of the
`BlockScore::Actions::Update` module to describe `:update` usage.