Remove erroneous force! test
This test was expecting retrieve().force! on a deleted candidate to return not found error. However the practice in this API is to retrieve any matching id from an object successfully AND if it was previously deleted to add a deleted: true attribute, rather than return a 404 Not Found. Hence this test is being removed In favor of testing the deleted attribute on the candidate.