Refactor RSpec arrangement and adjust namespaces
This commit is huge. For that I apologize. Details:

- Refactor the headers in the unit tests to combine the module BlockScore into the Spec describe block to not pollute the underlying namespace
- The above change has two effects 1) dedenting the code (making it all marked as changed), and 2) requiring a correct directory name block_score not blockscore
- Thus all the unit test moved to unit/block_score and now all the files are deleted and added
- This had the side effect of meaning that the cassettes were also regenerated in most cases under a new directory path
- The errors which were arbitrarily under an errors sub-directory are now under block_score directory
- the above necessitated changing the exclusions in rubocop.yml but otherwise satisfied naming conventions in rubocop.

In addition to these global changes, there are many internal changes:

- spec code was shortened to eliminate max line length faults (BTW that means that rubocop exclusions need to be revisited - they are not automatically dropped)
- all cases of ' eq ' tests where promoted to eql or equal to strengthen assertions
- all be_a comparisons were strengthened to an appropriate be_an_instance_of
- described_class was inserted everywhere appropriate

Also some tests from the various member classes where pushed into a new Base_spec.rb as that is where the unit tests should reside (nb: base_spec.rb is incomplete and will be addressed in short-term future)

What is not done?

The cases of be_persisted, etc. should be changed to persisted? be true to strengthen the assertion of predicate methods returning true or false.
And of course once this is in place than another pass to ensure method coverage.
Addition of specific unit tests for the actions is also a to-do.