Fix to reset the API_KEY each module tested
Because the suite now tests the API_KEY usage (null it out to prove it errors properly, etc.) depending on random usage sometime the API_KEY is not available for the rest of the test runs, causing massive errors. This resets the API_KEY at the beginning of each module.