Fix the vcr: true setup
Since VCR needs to be explicitly turned off, the prior approach of just running without VCR is insufficient. VCR needs to be "skipped" along with full external network access re-enabled if vcr: true is missing. This commit fixes that, so with vcr true then the recording mode is fetched from environment variable (as was prior). If vcr: true is missing, then the recording mode is set to :skip which triggers the enablement of the network and VCR is explicitly turned off for that example.